refactor: Split aggregates.rs (1034 lines) into focused test modules
Split `crates/executor/src/tests/aggregates.rs` (1,034 lines, 15 tests) into 5 focused modules:

- aggregate_count_sum_avg_tests.rs (408 lines, 7 tests) - COUNT, SUM, AVG with NULL handling
- aggregate_min_max_tests.rs (207 lines, 3 tests) - MIN, MAX on integers and strings
- aggregate_group_by_tests.rs (92 lines, 1 test) - GROUP BY with aggregates
- aggregate_having_tests.rs (95 lines, 1 test) - HAVING clause filtering
- aggregate_edge_case_tests.rs (236 lines, 3 tests) - Decimal precision, mixed types, CASE expressions

All 15 tests passing. Improves test organization and maintainability.
